+++ src/modules/Bot/Application/Service/Exchange/Exchange/InstrumentInfoDto.php
@@ -12,7 +12,7 @@ class InstrumentInfoDto implements JsonSerializable
         public float $minOrderQty,
         public float $minOrderValue,
         public float $minLeverage,
-        public float $maxLeverage,
+        public int $maxLeverage,
         public float $tickSize,
         public int $priceScale,
         public string $quoteCoin,
